Fix: make showcase default home page
