+++ src/pages/about/styles.scss
@@ -1,0 +1,5 @@
+@import '../../styles/base.scss';
+
+.home {
+    overflow: hidden;
+}
